pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Ruben Bimberg <ruben.bimberg@gmx.de>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "=3.11.10"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
pyjwt = "^2.9.0"
langchain = "^0.3.4"
langchain-openai = "^0.2.3"
langchain-anthropic = "^0.2.3"
langchain-mistralai = "^0.2.0"
langchain-fireworks = "^0.2.1"
langchain-together = "^0.2.0"
langchain-google-genai = "^2.0.1"
langchain-groq = "^0.2.0"
langchain-cohere = "^0.3.1"
langchain-aws = "^0.2.4"
langchain-huggingface = "^0.1.0"
langchain-nvidia-ai-endpoints = "^0.3.3"
langchain-ai21 = "^0.1.9"
langchain-upstage = "^0.3.0"
langchain-community = "^0.3.3"
langchain-ibm = "^0.3.2"
snowflake-snowpark-python = "^1.24.0"
asyncpg = "^0.30.0"
aiofiles = "^24.1.0"
python-multipart = "^0.0.17"
python-magic = "^0.4.27"
langchain-chroma = "^0.1.4"
imageio = "^2.36.0"
bs4 = "^0.0.2"
unstructured = {extras = ["all-docs"], version = "^0.16.4"}
pypandoc-binary = "^1.14"
more-itertools = "^10.5.0"
apscheduler = "^3.10.4"
langgraph = "^0.2.45"
psycopg = {extras = ["binary", "pool"], version = "^3.2.3"}
psycopg-pool = "^3.2.3"
langgraph-checkpoint-postgres = "^2.0.2"
torch = {version = "^2.5.1+cpu", source = "pytorch"}
torchvision = {version = "^0.20.1+cpu", source = "pytorch"}
torchaudio = {version = "^2.5.1+cpu", source = "pytorch"}

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
